+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/bookmark/BookmarkPlacesByRegionResponseDto.java
@@ -2,6 +2,7 @@ package ddd.caffeine.ratrip.module.place.presentation.dto.bookmark;
 
 import java.util.List;
 
+import ddd.caffeine.ratrip.module.place.domain.bookmark.repository.dao.BookmarkPlaceByRegionDao;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
